@@ -6,37 +6,26 @@ package com.bambooradical.monitor;
 import com.google.cloud.datastore.Datastore;
 import com.google.cloud.datastore.DatastoreOptions;
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.context.web.SpringBootServletInitializer;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.context.WebApplicationContext;
 
 /**
  * @created: 19/11/2016 22:11:12
  * @author : Peter Withers <peter@gthb-bambooradical.com>
  */
-@Configuration
-@ComponentScan
-@EnableAutoConfiguration
+
+@SpringBootApplication
 public class Application extends SpringBootServletInitializer {
 
     public static void main(String[] args) {
-        SpringApplication.run(applicationClass, args);
+        SpringApplication.run(Application.class, args);
     }
 
     @Override
     protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-        return application.sources(applicationClass);
-    }
-
-    private static Class<Application> applicationClass = Application.class;
-
-    @Override
-    protected WebApplicationContext run(SpringApplication application) {
-        return super.run(application);
+        return application.sources(Application.class);
     }
 
     @Bean
